refactor: features/mypage 폴더 구조를 탭 기준으로 그룹화 - common/: 왼쪽 사이드바 공통 컴포넌트 및 훅 (editable-profile-card, basic-info, report) - profile/: 프로필 탭 전용 컴포넌트 및 훅 (sections, dialogs, careers 등) - UserBasicInfo 타입을 entities/member/types로 이동
