--- src/features/mypage/types/index.ts
@@ -1,5 +1,0 @@
-import { FrontendStudentBasicInfo } from '@/entities/student';
-import { FrontendTeacherBasicInfo } from '@/entities/teacher';
-
-// TODO Parent 추가
-export type UserBasicInfo = FrontendTeacherBasicInfo | FrontendStudentBasicInfo;
